Add HttpProtocol enum and http_protocol field to Message
This commit adds a new `HttpProtocol` enum that represents HTTP versions
and adds an optional `http_protocol` field to the `Message` message.
This allows specifying the particular HTTP version used if the DOH
protocol is in use.

Hypothetically, this field could be used in the future to specify the
HTTP version in use for an HTTP-based DNS transport protocol that is not
the currently existing RFC 8484 DOH protocol. However, at the time of
this writing none of the other currently defined dnstap `SocketProtocol`
values are HTTP-based transports, so the `http_protocol` field
should only be set if `socket_protocol` is set to DOH. Otherwise, the
`http_protocol` field should be left unset.